split out icons from the help-sidebar
