+++ ui/src/app/components/help-sidebar.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { CSSProperties } from 'react';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
 import * as fp from 'lodash/fp';
 import { faEdit } from '@fortawesome/free-regular-svg-icons';
@@ -9,11 +8,9 @@ import {
   faFolderOpen,
   faInbox,
   faInfoCircle,
-  IconDefinition,
 } from '@fortawesome/free-solid-svg-icons';
 import { faCircle } from '@fortawesome/free-solid-svg-icons/faCircle';
 import { faDna } from '@fortawesome/free-solid-svg-icons/faDna';
-import { faLock } from '@fortawesome/free-solid-svg-icons/faLock';
 import { faSyncAlt } from '@fortawesome/free-solid-svg-icons/faSyncAlt';
 import { faTerminal } from '@fortawesome/free-solid-svg-icons/faTerminal';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -89,15 +86,10 @@ import times from 'assets/icons/times-light.svg';
 import moment from 'moment';
 
 import { RouteLink } from './app-router';
+import { displayRuntimeIcon, IconConfig, proIcons } from './help-sidebar-icons';
 import { RuntimeErrorModal } from './runtime-error-modal';
 
 export const LOCAL_STORAGE_KEY_SIDEBAR_STATE = 'WORKSPACE_SIDEBAR_STATE';
-
-const proIcons = {
-  arrowLeft: arrowLeft,
-  runtime: runtime,
-  times: times,
-};
 
 const styles = reactStyles({
   sidebarContainer: {
@@ -218,17 +210,6 @@ const iconStyles = {
 };
 
 export const LEONARDO_APP_PAGE_KEY = 'leonardo_app';
-
-interface IconConfig {
-  id: string;
-  disabled: boolean;
-  faIcon: IconDefinition;
-  label: string;
-  showIcon: () => boolean;
-  style: CSSProperties;
-  tooltip: string;
-  hasContent: true;
-}
 
 const pageKeyToAnalyticsLabels = {
   about: 'About Page',
@@ -581,127 +562,6 @@ export const HelpSidebar = fp.flow(
       );
     }
 
-    displayRuntimeIcon(icon: IconConfig) {
-      const { runtimeStore: store, compoundRuntimeOps, workspace } = this.props;
-      let status = store?.runtime?.status;
-      if (
-        (!status || status === RuntimeStatus.Deleted) &&
-        workspace.namespace in compoundRuntimeOps
-      ) {
-        // If a compound operation is still pending, and we're transitioning
-        // through the "Deleted" phase of the runtime, we want to keep showing
-        // an activity spinner. Avoids an awkward UX during a delete/create cycle.
-        // There also be some lag during the runtime creation flow between when
-        // the compound operation starts, and the runtime is set in the store; for
-        // this reason use Creating rather than Deleting here.
-        status = RuntimeStatus.Creating;
-      }
-
-      // We always want to show the thunderstorm icon.
-      // For most runtime statuses (Deleting and Unknown currently excepted), we will show a small
-      // overlay icon in the bottom right of the tab showing the runtime status.
-      return (
-        <FlexRow
-          style={{
-            height: '100%',
-            alignItems: 'center',
-            justifyContent: 'space-around',
-          }}
-        >
-          <img
-            data-test-id={'help-sidebar-icon-' + icon.id}
-            src={proIcons[icon.id]}
-            style={{ ...icon.style, position: 'absolute' }}
-          />
-          <FlexRow
-            data-test-id='runtime-status-icon-container'
-            style={styles.statusIconContainer}
-          >
-            {(() => {
-              const errIcon = (
-                <FontAwesomeIcon
-                  icon={faCircle}
-                  style={{
-                    ...styles.asyncOperationStatusIcon,
-                    ...styles.runtimeStatusIconOutline,
-                    color: colors.asyncOperationStatus.error,
-                  }}
-                />
-              );
-
-              if (store.loadingError) {
-                if (
-                  store.loadingError instanceof ComputeSecuritySuspendedError
-                ) {
-                  return (
-                    <FontAwesomeIcon
-                      icon={faLock}
-                      style={{
-                        ...styles.asyncOperationStatusIcon,
-                        color: colors.asyncOperationStatus.stopped,
-                      }}
-                    />
-                  );
-                }
-                return errIcon;
-              }
-              switch (status) {
-                case RuntimeStatus.Creating:
-                case RuntimeStatus.Starting:
-                case RuntimeStatus.Updating:
-                  return (
-                    <FontAwesomeIcon
-                      icon={faSyncAlt}
-                      style={{
-                        ...styles.asyncOperationStatusIcon,
-                        ...styles.rotate,
-                        color: colors.asyncOperationStatus.starting,
-                      }}
-                    />
-                  );
-                case RuntimeStatus.Stopped:
-                  return (
-                    <FontAwesomeIcon
-                      icon={faCircle}
-                      style={{
-                        ...styles.asyncOperationStatusIcon,
-                        ...styles.runtimeStatusIconOutline,
-                        color: colors.asyncOperationStatus.stopped,
-                      }}
-                    />
-                  );
-                case RuntimeStatus.Running:
-                  return (
-                    <FontAwesomeIcon
-                      icon={faCircle}
-                      style={{
-                        ...styles.asyncOperationStatusIcon,
-                        ...styles.runtimeStatusIconOutline,
-                        color: colors.asyncOperationStatus.running,
-                      }}
-                    />
-                  );
-                case RuntimeStatus.Stopping:
-                case RuntimeStatus.Deleting:
-                  return (
-                    <FontAwesomeIcon
-                      icon={faSyncAlt}
-                      style={{
-                        ...styles.asyncOperationStatusIcon,
-                        ...styles.rotate,
-                        color: colors.asyncOperationStatus.stopping,
-                      }}
-                    />
-                  );
-                case RuntimeStatus.Error:
-                  return errIcon;
-              }
-            })()}
-          </FlexRow>
-        </FlexRow>
-      );
-    }
-
     withinPastTwentyFourHours(epoch) {
       const completionTimeMoment = moment(epoch);
       const twentyFourHoursAgo = moment().subtract(1, 'days');
@@ -843,7 +703,7 @@ export const HelpSidebar = fp.flow(
             </a>
           ),
         ],
-        ['runtime', () => this.displayRuntimeIcon(icon)],
+        ['runtime', () => displayRuntimeIcon(icon, this.props)],
         [
           'terminal',
           () => (
